Splits GitActions into separate modules Removes more namespaces for better tree-shaking
