+++ src/quickpicks/items/commits.ts
@@ -1,10 +1,11 @@
 import type { QuickPickItem } from 'vscode';
 import { window } from 'vscode';
-import { GitActions } from '../../commands/gitCommands.actions';
 import type { OpenChangedFilesCommandArgs } from '../../commands/openChangedFiles';
 import { QuickCommandButtons } from '../../commands/quickCommand.buttons';
 import { Commands, GlyphChars } from '../../constants';
 import { Container } from '../../container';
+import { browseAtRevision } from '../../git/actions';
+import * as CommitActions from '../../git/actions/commit';
 import { CommitFormatter } from '../../git/formatters/commitFormatter';
 import type { GitCommit } from '../../git/models/commit';
 import type { GitFileChange } from '../../git/models/file';
@@ -76,7 +77,7 @@ export class CommitFileQuickPickItem extends CommandQuickPickItem {
 	}
 
 	override execute(options?: { preserveFocus?: boolean; preview?: boolean }): Promise<void> {
-		return GitActions.Commit.openChanges(this.file, this.commit, options);
+		return CommitActions.openChanges(this.file, this.commit, options);
 		// const fileCommit = await this.commit.getCommitForFile(this.file)!;
 
 		// if (fileCommit.previousSha === undefined) {
@@ -114,7 +115,7 @@ export class CommitBrowseRepositoryFromHereCommandQuickPickItem extends CommandQ
 	}
 
 	override execute(_options: { preserveFocus?: boolean; preview?: boolean }): Promise<void> {
-		return GitActions.browseAtRevision(this.commit.getGitUri(), {
+		return browseAtRevision(this.commit.getGitUri(), {
 			before: this.executeOptions?.before,
 			openInNewWindow: this.executeOptions?.openInNewWindow,
 		});
@@ -147,7 +148,7 @@ export class CommitCopyIdQuickPickItem extends CommandQuickPickItem {
 	}
 
 	override execute(): Promise<void> {
-		return GitActions.Commit.copyIdToClipboard(this.commit);
+		return CommitActions.copyIdToClipboard(this.commit);
 	}
 
 	override async onDidPressKey(key: Keys): Promise<void> {
@@ -162,7 +163,7 @@ export class CommitCopyMessageQuickPickItem extends CommandQuickPickItem {
 	}
 
 	override execute(): Promise<void> {
-		return GitActions.Commit.copyMessageToClipboard(this.commit);
+		return CommitActions.copyMessageToClipboard(this.commit);
 	}
 
 	override async onDidPressKey(key: Keys): Promise<void> {
@@ -179,7 +180,7 @@ export class CommitOpenAllChangesCommandQuickPickItem extends CommandQuickPickIt
 	}
 
 	override execute(options: { preserveFocus?: boolean; preview?: boolean }): Promise<void> {
-		return GitActions.Commit.openAllChanges(this.commit, options);
+		return CommitActions.openAllChanges(this.commit, options);
 	}
 }
 
@@ -189,7 +190,7 @@ export class CommitOpenAllChangesWithDiffToolCommandQuickPickItem extends Comman
 	}
 
 	override execute(): Promise<void> {
-		return GitActions.Commit.openAllChangesWithDiffTool(this.commit);
+		return CommitActions.openAllChangesWithDiffTool(this.commit);
 	}
 }
 
@@ -199,7 +200,7 @@ export class CommitOpenAllChangesWithWorkingCommandQuickPickItem extends Command
 	}
 
 	override execute(options: { preserveFocus?: boolean; preview?: boolean }): Promise<void> {
-		return GitActions.Commit.openAllChangesWithWorking(this.commit, options);
+		return CommitActions.openAllChangesWithWorking(this.commit, options);
 	}
 }
 
@@ -209,7 +210,7 @@ export class CommitOpenChangesCommandQuickPickItem extends CommandQuickPickItem 
 	}
 
 	override execute(options: { preserveFocus?: boolean; preview?: boolean }): Promise<void> {
-		return GitActions.Commit.openChanges(this.file, this.commit, options);
+		return CommitActions.openChanges(this.file, this.commit, options);
 	}
 }
 
@@ -219,7 +220,7 @@ export class CommitOpenChangesWithDiffToolCommandQuickPickItem extends CommandQu
 	}
 
 	override execute(): Promise<void> {
-		return GitActions.Commit.openChangesWithDiffTool(this.file, this.commit);
+		return CommitActions.openChangesWithDiffTool(this.file, this.commit);
 	}
 }
 
@@ -229,7 +230,7 @@ export class CommitOpenChangesWithWorkingCommandQuickPickItem extends CommandQui
 	}
 
 	override execute(options: { preserveFocus?: boolean; preview?: boolean }): Promise<void> {
-		return GitActions.Commit.openChangesWithWorking(this.file, this.commit, options);
+		return CommitActions.openChangesWithWorking(this.file, this.commit, options);
 	}
 }
 
@@ -239,7 +240,7 @@ export class CommitOpenDirectoryCompareCommandQuickPickItem extends CommandQuick
 	}
 
 	override execute(): Promise<void> {
-		return GitActions.Commit.openDirectoryCompareWithPrevious(this.commit);
+		return CommitActions.openDirectoryCompareWithPrevious(this.commit);
 	}
 }
 
@@ -249,7 +250,7 @@ export class CommitOpenDirectoryCompareWithWorkingCommandQuickPickItem extends C
 	}
 
 	override execute(): Promise<void> {
-		return GitActions.Commit.openDirectoryCompareWithWorking(this.commit);
+		return CommitActions.openDirectoryCompareWithWorking(this.commit);
 	}
 }
 
@@ -259,7 +260,7 @@ export class CommitOpenDetailsCommandQuickPickItem extends CommandQuickPickItem 
 	}
 
 	override execute(options: { preserveFocus?: boolean; preview?: boolean }): Promise<void> {
-		return GitActions.Commit.showDetailsView(this.commit, { preserveFocus: options?.preserveFocus });
+		return CommitActions.showDetailsView(this.commit, { preserveFocus: options?.preserveFocus });
 	}
 }
 
@@ -269,7 +270,7 @@ export class CommitOpenInGraphCommandQuickPickItem extends CommandQuickPickItem 
 	}
 
 	override execute(options: { preserveFocus?: boolean; preview?: boolean }): Promise<void> {
-		return GitActions.Commit.showInCommitGraph(this.commit, { preserveFocus: options?.preserveFocus });
+		return CommitActions.showInCommitGraph(this.commit, { preserveFocus: options?.preserveFocus });
 	}
 }
 
@@ -279,7 +280,7 @@ export class CommitOpenFilesCommandQuickPickItem extends CommandQuickPickItem {
 	}
 
 	override execute(_options: { preserveFocus?: boolean; preview?: boolean }): Promise<void> {
-		return GitActions.Commit.openFiles(this.commit);
+		return CommitActions.openFiles(this.commit);
 	}
 }
 
@@ -289,7 +290,7 @@ export class CommitOpenFileCommandQuickPickItem extends CommandQuickPickItem {
 	}
 
 	override execute(options?: { preserveFocus?: boolean; preview?: boolean }): Promise<void> {
-		return GitActions.Commit.openFile(this.file, this.commit, options);
+		return CommitActions.openFile(this.file, this.commit, options);
 	}
 }
 
@@ -299,7 +300,7 @@ export class CommitOpenRevisionsCommandQuickPickItem extends CommandQuickPickIte
 	}
 
 	override execute(_options: { preserveFocus?: boolean; preview?: boolean }): Promise<void> {
-		return GitActions.Commit.openFilesAtRevision(this.commit);
+		return CommitActions.openFilesAtRevision(this.commit);
 	}
 }
 
@@ -309,7 +310,7 @@ export class CommitOpenRevisionCommandQuickPickItem extends CommandQuickPickItem
 	}
 
 	override execute(options?: { preserveFocus?: boolean; preview?: boolean }): Promise<void> {
-		return GitActions.Commit.openFileAtRevision(this.file, this.commit, options);
+		return CommitActions.openFileAtRevision(this.file, this.commit, options);
 	}
 }
 
@@ -319,7 +320,7 @@ export class CommitApplyFileChangesCommandQuickPickItem extends CommandQuickPick
 	}
 
 	override async execute(): Promise<void> {
-		return GitActions.Commit.applyChanges(this.file, this.commit);
+		return CommitActions.applyChanges(this.file, this.commit);
 	}
 }
 
@@ -334,7 +335,7 @@ export class CommitRestoreFileChangesCommandQuickPickItem extends CommandQuickPi
 	}
 
 	override execute(): Promise<void> {
-		return GitActions.Commit.restoreFile(this.file, this.commit);
+		return CommitActions.restoreFile(this.file, this.commit);
 	}
 }
